move SelectedProfilesComponent into _shared module
