+++ src/app/report/report-parameter/_shared/selected-profiles/selected-profiles.component.scss
