+++ src/app/report/report.module.ts
@@ -40,7 +40,7 @@ import {UnitSearchComponent} from './report-parameter/report-a4-a6/unit-search/u
 import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
 import {FormsModule} from '@angular/forms';
 import {MatTableModule} from '@angular/material/table';
-import {SelectedProfilesComponent} from './report-parameter/report-a4-a6/selected-profiles/selected-profiles.component';
+import {SelectedProfilesComponent} from './report-parameter/_shared/selected-profiles/selected-profiles.component';
 import {FieldWrapperComponent} from './report-parameter/report-a4-a6/field-wrapper/field-wrapper.component';
 import {IssuerSearchComponent} from './report-parameter/report-a4-a6/issuer-search/issuer-search.component';
 import {MatPaginatorModule} from "@angular/material/paginator";
